+++ app/lib/shared/versions.dart
@@ -22,10 +22,10 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2022.08.02',
+  '2022.08.16',
   // Fallback runtime versions.
+  '2022.08.02',
   '2022.07.05',
-  '2022.06.23',
 ];
 
 /// Represents a combined version of the overall toolchain and processing,
@@ -57,7 +57,7 @@ final String runtimeSdkVersion = '2.17.0';
 final String toolStableDartSdkVersion = '2.17.6';
 final String toolStableFlutterSdkVersion = '3.0.5';
 final String toolPreviewDartSdkVersion = '2.18.0-271.4.beta';
-final String toolPreviewFlutterSdkVersion = '3.3.0-0.1.pre';
+final String toolPreviewFlutterSdkVersion = '3.3.0-0.3.pre';
 
 final semanticToolStableDartSdkVersion =
     Version.parse(toolStableDartSdkVersion);