+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bumped runtimeVersion to `2022.08.16`.
+ * Upgraded preview Flutter analysis SDK to `3.3.0-0.3.pre`.
 
 ## `20220804t094500-all`
  * Bumped runtimeVersion to `2022.08.02`.